+++ src/components/profile/NewTaste/NewTasteCategoryListItem.tsx
