+++ src/pages/profile/EditProfile/EditProfile.tsx
@@ -2,8 +2,8 @@ import styled from '@emotion/styled'
 import { useRecoilValue } from 'recoil'
 import Dimmer from '@components/layouts/Dimmer'
 import { FlexBox } from '@components/layouts/FlexBox'
-import EditProfileImage from '@components/profile/EditProfileImage'
-import ProfileMenuButtons from '@components/profile/ProfileMenuButtons'
+import EditProfileImage from '@components/profile/EditProfile/EditProfileImage'
+import ProfileMenuButtons from '@components/profile/ProfileInfo/ProfileMenuButtons'
 import EditUserInfo from './EditUserInfo'
 import { profileState } from '@libs/store/profile'
 import { useOutsideClick } from '@libs/hooks/useOutsideClick'
