+++ src/pages/profile/EditProfile/EditUserInfo.tsx
@@ -7,8 +7,8 @@ import { Birth } from '@pages/onboarding/details/signup/Birth'
 import { Spacing } from '@components/atoms/Spacing'
 import { FlexBox } from '@components/layouts/FlexBox'
 import { Padding } from '@components/layouts/Padding'
-import OnboardingStatus from '@components/profile/OnboardingStatus'
-import EditProfileButton from './EditProfileButton'
+import OnboardingStatus from '@components/profile/EditProfile/OnboardingStatus'
+import EditProfileButton from '../../../components/profile/EditProfile/EditProfileButton'
 import { authState } from '@libs/store/auth'
 import { onboardingState } from '@libs/store/onboard'
 import { profileState } from '@libs/store/profile'
