+++ src/pages/profile/NewTaste.tsx
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query'
 import styled from '@emotion/styled'
 import { Padding } from '@components/layouts/Padding'
 import { Spacing } from '@components/atoms/Spacing'
-import NewTasteCategory from '@components/profile/NewTasteCategory'
+import NewTasteCategory from '@components/profile/NewTaste/NewTasteCategory'
 import { authState } from '@libs/store/auth'
 import { UserApi } from '@utils/apis/user/UserApi'
 import { UserNewTasteCategory } from '@utils/apis/user/UserType'
