+++ src/pages/profile/PastToday/PastTodayDetail.tsx
