+++ src/pages/profile/ShareProfile.tsx
@@ -1,6 +1,6 @@
 import { Spacing } from '@components/atoms/Spacing'
-import ShareProfileButtons from '@components/profile/ShareProfileButtons'
-import ShareProfileImage from '@components/profile/ShareProfileImage'
+import ShareProfileButtons from '@components/profile/ShareProfile/ShareProfileButtons'
+import ShareProfileImage from '@components/profile/ShareProfile/ShareProfileImage'
 import styled from '@emotion/styled'
 
 const ShareProfile = () => {
