+++ src/pages/profile/index.tsx
@@ -20,8 +20,8 @@ import HCCUP from '@pages/searchTaste/HCCUP'
 import AppBarTemplate from '@components/layouts/AppBarTemplate'
 import Loading from '@components/atoms/Loading'
 import { useNavigate } from 'react-router-dom'
-import EditProfile from './EditProfile'
-import EditOnboardingStatus from './EditOnboardingStatus'
+import EditProfile from './EditProfile/EditProfile'
+import EditOnboardingStatus from './EditProfile/EditOnboardingStatus'
 import Profile from './Profile'
 import NewTaste from './NewTaste'
 import { authState } from '@libs/store/auth'
@@ -29,12 +29,12 @@ import MenuIcon from '@assets/icons/MenuIcon'
 import { useRecoilState } from 'recoil'
 import { answerState } from '@libs/store/question'
 import { UserApi } from '@utils/apis/user/UserApi'
-import PastToday from './PastToday'
-import PastTodayDetail from './PastTodayDetail'
+import PastToday from './PastToday/PastToday'
+import PastTodayDetail from './PastToday/PastTodayDetail'
 import FullModal from '@components/atoms/Modal/FullModal'
 import BigX from '@assets/icons/BigX'
 import ShareProfile from './ShareProfile'
-import EditFavorBox from '@pages/profile/EditFavorBox'
+import EditFavorBox from './EditfavorBox/EditFavorBox'
 
 const ProfileRouter = () => {
   const auth = useRecoilValue(authState)
